third_party/WebKit/LayoutTests/media/autoplay-non-whitelisted-scope.html

<!DOCTYPE html>
<title>Test for autoplay of whitelisted frames</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="media-controls.js"></script>
<body>
<script>

test(function() {
  assert_true(!!window.internals
      && !!internals.settings,
              "This test only works when run as a layout test!");
}, "Prerequisites to running the rest of the tests");

internals.settings.setAutoplayPolicy('user-gesture-required');
internals.settings.setMediaPlaybackGestureWhitelistScope(
    document.URL.substring(0, document.URL.lastIndexOf('/') + 1) + "foo/");

testRunner.setAutoplayAllowed(true);

function createMediaElement(type) {
  var e = document.createElement(type);
  e.src = type == 'audio' ? 'content/test.oga' : 'content/test.ogv';
  return e;
}

function createVideoElement() {
  return createMediaElement('video');
}

function createAudioElement() {
  return createMediaElement('audio');
}

async_test(function(t) {
  var e = createVideoElement();
  e.autoplay = true;
  document.body.appendChild(e);

  var expectedEvents = [ 'canplay' ];
  var eventWatcher = new EventWatcher(t, e, expectedEvents);
  eventWatcher.wait_for(expectedEvents).then(
      t.step_func_done(function() {
        assert_true(e.paused);
      }));
}, "Test that a video with an autoplay attribute doesn't autoplay when not whitelisted.");

promise_test(function(t) {
  return promise_rejects(
      t,
      new DOMException(
          'play() can only be initiated by a user gesture.',
          'NotAllowedError'),
      createVideoElement().play());
}, "Test that play() on a video fails without gesture when not whitelisted.");

async_test(function(t) {
  var e = createAudioElement();
  e.autoplay = true;
  document.body.appendChild(e);

  var expectedEvents = [ 'canplay' ];
  var eventWatcher = new EventWatcher(t, e, expectedEvents);
  eventWatcher.wait_for(expectedEvents).then(
      t.step_func_done(function() {
        assert_true(e.paused);
      }));
}, "Test that an audio with an autoplay attribute doesn't autoplay when not whitelisted.");

promise_test(function(t) {
  return promise_rejects(
      t,
      new DOMException(
          'play() can only be initiated by a user gesture.',
          'NotAllowedError'),
      createAudioElement().play());
}, "Test that play() on an audio fails without gesture when not whitelisted.");

</script>
</body>
